Convert all conversions to ESM and rename them
