+++ src/conversions/cielab.js
@@ -1,11 +1,11 @@
-const getIlluminant = require('../helpers/get-illuminant.js').default
+import getIlluminant from '../helpers/get-illuminant'
 
 const epsilon = 0.008856
 const kappa = 903.3
 const white = getIlluminant('D65')
 const toR = f => Math.pow(f, 3) > epsilon ? Math.pow(f, 3) : ((116 * f) - 16) / kappa
 
-module.exports = {
+export default {
   XYZ: value => {
     const Fy = (value.L + 16) / 116
     const Fx = (value.a / 500) + Fy
