+++ src/conversions/cielch.js
@@ -1,20 +1,33 @@
-const getTransform = require('../helpers/get-transform.js').default
-const toRadian = require('../helpers/to-radian.js').default
+import toRadian from '../helpers/to-radian'
+import getTransform from '../helpers/get-transform'
 
-module.exports = {
-  cielch: value => {
+export default {
+  cieluv: value => {
+    const h = toRadian(value.h)
+
+    const u = value.C * Math.cos(h)
+    const v = value.C * Math.sin(h)
+
+    return {
+      L: value.L,
+      u,
+      v
+    }
+  },
+
+  hsluv: value => {
     if (value.L > 99.9999999) {
-      return { L: 100, C: 0, h: value.hu }
+      return { hu: value.h, s: 0, l: 100 }
     }
     if (value.L < 0.00000001) {
-      return { L: 0, C: 0, h: value.hu }
+      return { hu: value.h, s: 0, l: 0 }
     }
 
     const epsilon = 0.008856
     const kappa = 903.3
 
-    const s1 = (value.l + 16) / 1560896
-    const s2 = s1 > epsilon ? s1 : value.l / kappa
+    const s1 = (value.L + 16) / 1560896
+    const s2 = s1 > epsilon ? s1 : value.L / kappa
 
     const m = getTransform('INVERSE_SRGB_XYZ')
     let rays = []
@@ -26,7 +39,7 @@ module.exports = {
 
       for (let t = 0; t < 2; t++) {
         let top1 = (284517 * m1 - 94839 * m3) * s2
-        let top2 = (838422 * m3 + 769860 * m2 + 731718 * m1) * value.l * s2 - 769860 * t * value.l
+        let top2 = (838422 * m3 + 769860 * m2 + 731718 * m1) * value.L * s2 - 769860 * t * value.L
         let bottom = (632260 * m3 - 126452 * m2) * s2 + 126452 * t
 
         rays.push({
@@ -37,7 +50,7 @@ module.exports = {
     }
 
     var min = Number.MAX_VALUE
-    let hrad = toRadian(value.hu)
+    let hrad = toRadian(value.h)
 
     rays.forEach((ray) => {
       let length = ray.b / (Math.sin(hrad) - ray.m * Math.cos(hrad))
@@ -49,9 +62,9 @@ module.exports = {
     let max = min
 
     return {
-      L: value.l,
-      C: max / 100 * value.s,
-      h: value.hu
+      hu: value.h,
+      s: value.C / max * 100,
+      l: value.L
     }
   }
 }
